refactor: gameTable을 컨트롤러로 이동
